♻️ Move parser traits into  `parser` module
This way `lib.rs` only contains module declarations and re-exports. Parser module documentation was provided by Claude Sonet 4.6.
