+++ src/combinators/attempt.rs
@@ -75,7 +75,6 @@ where
 
 #[cfg(test)]
 mod tests {
-	use crate::input::position::Position;
 	use crate::*;
 
 	#[test]
