+++ src/combinators/choice.rs
@@ -49,7 +49,6 @@ where
 
 #[cfg(test)]
 mod tests {
-	use crate::input::position::Position;
 	use crate::*;
 
 	#[test]
