+++ src/combinators/look_ahead.rs
@@ -61,7 +61,6 @@ where
 
 #[cfg(test)]
 mod tests {
-	use crate::input::position::Position;
 	use crate::*;
 
 	#[test]
