+++ src/combinators/many_until.rs
@@ -44,7 +44,6 @@ where
 
 #[cfg(test)]
 mod tests {
-	use crate::input::position::Position;
 	use crate::*;
 
 	#[test]
