+++ src/combinators/option.rs
@@ -55,7 +55,6 @@ where
 
 #[cfg(test)]
 mod tests {
-	use crate::input::position::Position;
 	use crate::*;
 
 	#[test]
