+++ src/combinators/satisfy.rs
@@ -59,7 +59,6 @@ where
 
 #[cfg(test)]
 mod tests {
-	use crate::input::position::Position;
 	use crate::*;
 
 	#[test]
